Fix CUDA OOM in add_gumbel_noise by using float32 instead of float64
Every temperature>0 run in Task4 crashed with the exact same OOM at this
line (confirmed via infer logs: all 9 temperature=0.3 combos failed with
identical traceback, not just combo 11). Root cause: float64 upcast of
the full (batch, seq_len, vocab=155136) logits tensor plus a same-size
noise tensor pushes past the ~1.5GB headroom left after the model itself
already uses ~22GB on a 24GB card. float32 halves that overhead.

+++ generate.py
@@ -14,8 +14,10 @@ def add_gumbel_noise(logits, temperature):
     '''
     if temperature == 0:
         return logits
-    logits = logits.to(torch.float64)
-    noise = torch.rand_like(logits, dtype=torch.float64)
+    # float64 here reliably OOMs on 24GB cards at this vocab size (155136) once the
+    # base model + activations already use ~22GB -- float32 halves that overhead.
+    logits = logits.to(torch.float32)
+    noise = torch.rand_like(logits, dtype=torch.float32)
     gumbel_noise = (- torch.log(noise)) ** temperature
     return logits.exp() / gumbel_noise
